refactor: move shared secret types into binstalk-types
