+++ crates/bin/Cargo.toml
@@ -27,6 +27,7 @@ pkg-fmt = "zip"
 atomic-file-install = { version = "1.0.14", path = "../atomic-file-install" }
 binstalk = { path = "../binstalk", version = "0.28.71", default-features = false }
 binstalk-manifests = { path = "../binstalk-manifests", version = "0.17.30" }
+binstalk-types = { path = "../binstalk-types", version = "0.10.6" }
 clap = { version = "4.5.54", features = ["derive", "env", "wrap_help"] }
 clap-cargo = "0.18.0"
 clap-markdown = { version = "0.1.5", optional = true }
