+++ crates/binstalk-manifests/src/helpers.rs
@@ -1,4 +1,4 @@
-use std::{fmt, fs, io, ops::Deref, path::Path};
+use std::{fs, io, path::Path};
 
 use fs_lock::FileLock;
 
@@ -12,28 +12,4 @@ pub(crate) fn create_if_not_exist(path: &Path) -> io::Result<FileLock> {
         .open(path)
         .and_then(FileLock::new_exclusive)
         .map(|file_lock| file_lock.set_file_path(path))
-}
-
-#[repr(transparent)]
-#[derive(Clone, Default)]
-pub struct Redacted<T>(T);
-
-impl<T> Redacted<T> {
-    pub const fn new(value: T) -> Self {
-        Self(value)
-    }
-}
-
-impl<T> Deref for Redacted<T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl<T> fmt::Debug for Redacted<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str("<redacted>")
-    }
 }
