+++ crates/binstalk-registry/Cargo.toml
@@ -16,7 +16,6 @@ base16 = "0.2.1"
 binstalk-downloader = { version = "0.13.37", path = "../binstalk-downloader", default-features = false, features = [
     "json",
 ] }
-binstalk-manifests = { version = "0.17.30", path = "../binstalk-manifests" }
 binstalk-types = { version = "0.10.6", path = "../binstalk-types" }
 cargo-toml-workspace = { version = "7.0.9", path = "../cargo-toml-workspace" }
 compact_str = { version = "0.9.0", features = ["serde"] }
@@ -37,7 +36,6 @@ tokio = { version = "1.51.0", features = [
 ], default-features = false }
 tracing = "0.1.43"
 url = "2.5.8"
-zeroize = "1.8.1"
 
 [dev-dependencies]
 tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
