+++ crates/binstalk-types/Cargo.toml
@@ -21,6 +21,7 @@ strum = "0.28.0"
 strum_macros = "0.28.0"
 target-lexicon = { version = "0.13.0", features = ["std"] }
 url = { version = "2.5.8", features = ["serde"] }
+zeroize = "1.8.1"
 
 [dev-dependencies]
 serde_json = "1"
