+++ crates/binstalk-types/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod cargo_toml_binstall;
 pub mod crate_info;
+pub mod secrets;
 
 pub use maybe_owned;
+pub use secrets::{Redacted, SecretString};
